+++ CHANGELOG.md
@@ -27,6 +27,11 @@ This is an example of a brief overview of the _Major_ or _Minor_ version changes
 
 ---
 
+## [6.5] border focus and transition
+
+- Added `border--focus` and `border--focus--noTransition' helper classes for standard focus styles
+- Added `transition-default` helper class for default transition
+
 ## [6.4] `Avatar`
 
 - Adds `Avatar` component for displaying avatar initials and images